refactor org domains section
